Выносит функции отрисовки в отдельный файл
